feat: remove env package and using only config package for configuration
The `env` package was just a hacky way for getting around the cyclic
import because the `config` package was using the `asserts` package
which imports the `pretty-logger` package and that package imports the
`config` package to determined what logger to return at runtime creating
a cyclic dependency.

Now instead using the `asserts` package in the `config` package it's
simlpy checking for errors and call `log.Fatalln` if something gose
wrong.